pyproject.toml
[project]
name = "genpulse"
version = "0.1.0"
description = "High-concurrency generative AI backend system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "asyncpg>=0.31.0",
    "boto3>=1.42.30",
    "click>=8.3.1",
    "comfy-cli>=1.3.8",
    "diffusers>=0.36.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "torch>=2.9.1",
    "transformers>=4.57.6",
    "tencentcloud-sdk-python-vod>=3.0.1340",
    "baidubce>=0.8.99",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
    "pyyaml>=6.0.2",
]

[project.scripts]
genpulse = "genpulse.cli:cli"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["src"]
testpaths = ["tests"]
